requirements.txt
Flask
requests
flask-cors
gunicorn
python-telegram-bot  # 👈 هذا هو الإضافة الضرورية